Mock function declaration macros is removed.
The functions are be already declared in the header files of the
project. So it makes no sense to declare the same functions again for
fake functions. This change thus forces the inclusion of the project
header files and so the original declaration, will be used for fake
functions as well.