Disconnect Global & Ordered expectation matches
Previously if the same expectation was added to both ordered and global
lists then if a function call satisfied expectations from both lists,
they both were marked as satisfied. This caused some results which might
not be desirable always. For example

```c
void test()
{
    fnA();
    fnC();
}

YT_TEST (simple, test_1)
{
    YT_MUST_CALL_ANY_ORDER(fnC);
    YT_MUST_CALL_IN_ORDER (fnA);
    YT_MUST_CALL_IN_ORDER (fnC);

    test();

    YT_END();
}
```

You see there are two expectations on `fnC` that one must be called
after `fnA` and another at any time. So one might assume that there must
at least be two calls to `fnC`, but that was not so, it would also
satisfy the `test` function you see above. It worked like this. There
are three expectations, one in the global list (from ANY_ORDER macro)
and two in the ordered list. In the `test` function call to `fnA` would
satisfy the ordered `fnA` expectation (which would be first in the
ordered list), then the call to `fnC` would match both expectations of
`fnC` (one in the ordered list & another in the ordered list). Thus this
would satisfy all the expectations.

Note that in one way of thinking this behaviour is not wrong, but may be
confusing at times. However there is another problem, the same
expectations would also match if there were two `fnC` calls like so

```c
void test()
{
    fnC();
    fnA();
    fnC();
}
```

This is what I meant when I said confusing. In the new behaviour, each
expectation is taken separately. This is done here like this, we first
match to see if the current call matches any global expectation, if it
does then we mark only that expectation as satisfied and nothing else,
even though the same would have also matched the expectation that at the
top of the ordered list. This means that the above expectations would
require at-least two calls to `fnC`, one at anytime and another after
`fnA` call to satisfy (which was the original idea)

This means the following would not pass our expectations, while it did
previously.

```c
void test()
{
    fnA();
    fnC();
}
```

But these would.

```c
void test()
{
    fnA();
    fnC();
    fnC();
}
```

```c
void test()
{
    fnC();
    fnA();
    fnC();
}
```

and so on.