Fix: Test parameter names & types in wrong order
This bug caused function parameters of parameterised tests to have wrong
types. The first parameter which has name '_a' had type specified for
the last parameter, and the last parameter had type specified for the
first parameter.

So YT_TESTP(foo, example, int, char) produced the following expansions
(see that the test parameters have wrong type as mentioned above).

```
    void example_test(size_t count, size_t test_number, char _a[], int _b[]) {
        ...
    }

    void example(size_t count, char _a[], int _b[]) {
        ...
    }
```

The correct expansion is

```
    void example_test(size_t count, size_t test_number, int _a[], char _b[]) {
        ...
    }

    void example(size_t count, int _a[], char _b[]) {
        ...
    }
```

The solution was to start the generation of argument and parameter
expansion from the left and stop when there no more arguments to the
right.

Also includes:
* Example which hits this bug